requirements.txt
accelerate==1.0.1
datasets==3.2.0
deepspeed==0.16.4
evaluate==0.4.0
flash-attn==2.7.2.post1
numpy==1.26.4
pandas==2.2.3
PyYAML==6.0.2
sacrebleu==2.4.3
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
soundfile==0.12.1
tensorboardX==2.6.2.2
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.5
transformers==4.49.0
